+++ docs/deployment/schedulers/scheduler-management.md
@@ -122,7 +122,7 @@ At this time, the following dokku commands are used to implement a complete sche
 - `apps:clone`: handles app cloning
     - triggers: post-app-clone-setup
 - `cron`: generates cron tasks for the app
-    - triggers: scheduler-cron-write
+    - triggers: scheduler-uses-host-cron, scheduler-cron-write
 - `deploy`: deploys app proceses and checks the status of a deploy
     - triggers: scheduler-app-status, scheduler-deploy, scheduler-is-deployed, scheduler-logs-failed
 - `enter`: enters a running container
