+++ plugins/scheduler-docker-local/.gitignore
@@ -3,6 +3,5 @@
 /cron-*
 /report
 /report-subcommand
-/scheduler-cron-write
 /scheduler-storage-exec
 /subcommands/report
