--- plugins/scheduler-docker-local/post-deploy
@@ -1,15 +1,0 @@
-#!/usr/bin/env bash
-set -eo pipefail
-source "$PLUGIN_CORE_AVAILABLE_PATH/common/functions"
-[[ $DOKKU_TRACE ]] && set -x
-
-trigger-scheduler-docker-local-post-deploy() {
-  declare desc="scheduler-docker-local post-deploy plugin trigger"
-  declare trigger="post-deploy"
-  declare APP="$1"
-
-  local DOKKU_SCHEDULER="$(get_app_scheduler "$APP")"
-  plugn trigger scheduler-cron-write "$DOKKU_SCHEDULER" "$APP"
-}
-
-trigger-scheduler-docker-local-post-deploy "$@"
