+++ plugins/scheduler-docker-local/scheduler-stop
@@ -50,8 +50,6 @@ trigger-scheduler-docker-local-scheduler-stop() {
       "$DOCKER_BIN" container rm --force $DOKKU_APP_CIDS &>/dev/null || true
     fi
   fi
-
-  plugn trigger scheduler-cron-write "$DOKKU_SCHEDULER" "$APP"
 }
 
 trigger-scheduler-docker-local-scheduler-stop "$@"
