--- plugins/scheduler-docker-local/triggers.go
@@ -1,6 +1,0 @@
-package schedulerdockerlocal
-
-// TriggerSchedulerCronWrite force updates the cron file for all apps
-func TriggerSchedulerCronWrite(scheduler string) error {
-	return writeCronTab(scheduler)
-}
